Memory_Service/requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-PyMongo==2.3.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
psutil==5.8.0
py-healthcheck==1.10.1
pymongo==3.11.3
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
urllib3==1.26.4
Werkzeug==1.0.1
prometheus_flask_exporter